+++ E.Deezer/Api/Genre.cs
@@ -11,7 +11,7 @@ namespace E.Deezer.Api
 {
     public interface IGenre : IObjectWithImage
     {
-        int Id { get; set; }
+        ulong Id { get; set; }
         string Name { get; set; }
 
         //METHODS
@@ -56,7 +56,7 @@ namespace E.Deezer.Api
 
     internal class Genre : ObjectWithImage, IGenre, IDeserializable<DeezerClient>
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public string Name { get; set; }
 
         public DeezerClient Client { get; set; }
